Rename source directory to same as package name
Files are unchanged other than the directory rename.
Git diffs present in this commit can be safely ignored.
