+++ m1wengine/dict_structures/animation_dict.py
