+++ m1wengine/enums/eaten_powers.py
