+++ m1wengine/file_managers/sprite_sheet.py
