+++ m1wengine/managers/camera_manager.py
