+++ m1wengine/tiles/entities/entity.py
